refactor: use task enum
