+++ data/image_dataset.py
@@ -11,7 +11,7 @@ from monai.data.meta_obj import set_track_meta
 from multiprocessing import cpu_count
 from math import ceil
 
-from utils.metrics import Task
+from utils.enums import Task
 from data.unalignedZipDataset import UnalignedZipDataset
 from utils.enums import Phase
 set_track_meta(False)
@@ -76,8 +76,6 @@ def get_dataset(config: dict[str, dict], phase: str, batch_size=None) -> DataLoa
             data_set = UnalignedZipDataset(data, transform, phase, config["General"]["inference"])
             loader = DataLoader(data_set, batch_size=batch_size or config[phase].get("batch_size") or 1, shuffle=phase!=Phase.TEST, num_workers=8, pin_memory=torch.cuda.is_available())
             return loader
-
-
 
     data_set = Dataset(train_files, transform=transform)
     loader = DataLoader(data_set, batch_size=batch_size or config[phase].get("batch_size") or 1, shuffle=phase!=Phase.TEST, num_workers=ceil(cpu_count()/2), pin_memory=torch.cuda.is_available())
